# api/openapi.yaml
openapi: 3.0.3
info:
  title: WalletHub API
  description: |
    WalletHub - платёжный шлюз для управления кошельками и транзакциями.

    ## Основные возможности
    - Управление пользователями и KYC верификация
    - Создание и управление мультивалютными кошельками (Fiat + Crypto)
    - Проведение финансовых операций (deposit, withdraw, transfer)
    - Полная история транзакций

    ## Аутентификация
    API использует Bearer Token аутентификацию. Передайте токен в заголовке:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Идемпотентность
    Все финансовые операции поддерживают идемпотентность через `idempotency_key`.
    При повторном запросе с тем же ключом вернётся результат первой операции.

    ## Rate Limiting
    - Общий лимит: 100 запросов/минуту
    - Финансовые операции: 30 запросов/минуту

    Заголовки ответа:
    - `X-RateLimit-Limit`: Максимум запросов
    - `X-RateLimit-Remaining`: Оставшееся количество
    - `X-RateLimit-Reset`: Unix timestamp сброса

    ## Формат ответа
    Все ответы имеют единый формат:
    ```json
    {
      "success": true/false,
      "data": { ... },
      "error": { "code": "...", "message": "..." },
      "meta": { "page": 1, "total": 100 },
      "request_id": "uuid",
      "timestamp": "2024-01-01T00:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: WalletHub Support
    email: support@wallethub.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /api/v1/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      description: Get paginated list of transactions with optional filters
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: wallet_id
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
  /api/v1/transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction by ID
      description: Get transaction details by UUID
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ApiError'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        details:
          type: object
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        idempotency_key:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          $ref: '#/components/schemas/TransactionStatus'
        amount:
          type: string
        currency_code:
          type: string
        destination_wallet_id:
          type: string
          format: uuid
          nullable: true
        external_reference:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        failure_reason:
          type: string
        retry_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    TransactionType:
      type: string
      enum: [DEPOSIT, WITHDRAW, PAYOUT, TRANSFER, FEE, REFUND, ADJUSTMENT]
    TransactionStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Transaction'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            transactions:
              type: array
              items:
                $ref: '#/components/schemas/Transaction'
            total_count:
              type: integer
        meta:
          $ref: '#/components/schemas/ApiMeta'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time